Move tests from pgmodel to pkg
