+++ .github/workflows/go-scheduled.yml
@@ -38,13 +38,13 @@ jobs:
       run: go test -v -race ./...
 
     - name: Test W/O Promscale Extension
-      run: go test -v -race ./pkg/pgmodel/end_to_end_tests/ -use-extension=false
+      run: go test -v -race ./pkg/tests/end_to_end_tests/ -use-extension=false
 
     - name: Test W/O TimescaleDB Extension
-      run: go test -v -race ./pkg/pgmodel/end_to_end_tests/ -use-extension=false -use-timescaledb=false
+      run: go test -v -race ./pkg/tests/end_to_end_tests/ -use-extension=false -use-timescaledb=false
 
     - name: Test extended dataset
-      run: go test -v -race -timeout=30m ./pkg/pgmodel/end_to_end_tests/ -extended-test
+      run: go test -v -race -timeout=30m ./pkg/tests/end_to_end_tests/ -extended-test
 
     - name: Generated
       run: |
