+++ Makefile
@@ -5,11 +5,12 @@ test:
 	go test -v -race ./...
 
 e2e-test:
-	go test -v ./pkg/pgmodel/end_to_end_tests/ -use-extension=false
-	go test -v ./pkg/pgmodel/end_to_end_tests/ -use-extension=false go test -v ./pkg/pgmodel/end_to_end_tests/ -use-extension=false -use-timescaledb=false
-	go test -v ./pkg/pgmodel/end_to_end_tests/ -use-timescale2
-	go test -v ./pkg/pgmodel/end_to_end_tests/ -use-extension=false -use-timescale2
-	go test -v ./pkg/pgmodel/end_to_end_tests/ -use-multinode
+	go test -v ./pkg/tests/end_to_end_tests/ -use-extension=false
+	go test -v ./pkg/tests/end_to_end_tests/ -use-extension=false
+	go test -v ./pkg/tests/end_to_end_tests/ -use-extension=false -use-timescaledb=false
+	go test -v ./pkg/tests/end_to_end_tests/ -use-timescale2
+	go test -v ./pkg/tests/end_to_end_tests/ -use-extension=false -use-timescale2
+	go test -v ./pkg/tests/end_to_end_tests/ -use-multinode
 
 
 go-fmt:
