+++ pkg/tests/end_to_end_tests/delete_test.go
