+++ pkg/tests/end_to_end_tests/drop_test.go
