+++ pkg/tests/end_to_end_tests/golden_files_test.go
