+++ pkg/tests/end_to_end_tests/nan_test.go
