+++ pkg/tests/end_to_end_tests/view_test.go
