+++ pkg/tests/test_migrations/sql/preinstall/001-setup.sql
