+++ pkg/tests/testdata/sql/info_view.sql
